@@ -1,6 +1,6 @@
 {
   "name": "@3kles/kles-material-table",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "scripts": {
     "build:lib": "ng build kles-material-table --prod",
     "publish:lib": "cd dist/kles-material-table && npm publish --registry 'http://192.168.111.61:8081/repository/npm-3kles/'  --force",
@@ -15,54 +15,54 @@
   "dependencies": {
     "@3kles/kles-material-dialog": "^1.0.0",
     "@3kles/kles-material-dynamicforms": "^1.1.10",
-    "@angular/animations": "^9.1.13",
-    "@angular/cdk": "^9.0.0",
-    "@angular/common": "^9.1.13",
-    "@angular/compiler": "^9.1.13",
-    "@angular/core": "^9.1.13",
+    "@angular/animations": "~12.0.1",
+    "@angular/cdk": "^12.0.1",
+    "@angular/common": "~12.0.1",
+    "@angular/compiler": "~12.0.1",
+    "@angular/core": "~12.0.1",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "^9.1.13",
-    "@angular/material": "^9.0.0",
-    "@angular/material-moment-adapter": "^9.0.0",
-    "@angular/platform-browser": "^9.1.13",
-    "@angular/platform-browser-dynamic": "^9.1.13",
-    "@angular/router": "^9.1.13",
+    "@angular/forms": "~12.0.1",
+    "@angular/material": "^12.0.1",
+    "@angular/material-moment-adapter": "^12.0.1",
+    "@angular/platform-browser": "~12.0.1",
+    "@angular/platform-browser-dynamic": "~12.0.1",
+    "@angular/router": "~12.0.1",
     "@infor-up/m3-odin": "^4.0.0",
-    "@ngx-translate/core": "^12.1.2",
+    "@ngx-translate/core": "^12.1.1",
     "@ngx-translate/http-loader": "^4.0.0",
-    "fp-ts": "^1.13.0",
+    "fp-ts": "^2.11.5",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "ngx-color-picker": "^10.1.0",
+    "ngx-color-picker": "^11.0.0",
     "ngx-mat-select-search": "^3.3.0",
-    "polytype": "^0.11.0",
+    "polytype": "^0.13.2",
     "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "xlsx": "^0.17.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.15",
+    "@angular-devkit/build-angular": "^12.2.16",
     "@angular-devkit/build-ng-packagr": "^0.901.15",
-    "@angular/cli": "^9.1.15",
-    "@angular/compiler-cli": "^9.1.13",
-    "@angular/language-service": "^9.1.13",
-    "@types/jasmine": "~3.5.0",
+    "@angular/cli": "~12.0.1",
+    "@angular/compiler-cli": "~12.0.1",
+    "@angular/language-service": "~12.0.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.168",
-    "@types/node": "^12.20.11",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.4.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^9.0.0",
-    "protractor": "~5.4.3",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^12.2.6",
+    "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.2.4"
   }
 }
