@@ -1,31 +1,31 @@
 {
   "name": "@3kles/kles-material-table",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "peerDependencies": {
     "@3kles/kles-material-dynamicforms": "^1.1.7",
-    "@angular/animations": "~9.1.0",
-    "@angular/cdk": "^9.0.0",
-    "@angular/common": "~9.1.0",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "~9.1.0",
+    "@angular/animations": "~12.0.1",
+    "@angular/cdk": "^12.0.1",
+    "@angular/common": "~12.0.1",
+    "@angular/compiler": "~12.0.1",
+    "@angular/core": "~12.0.1",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "~9.1.0",
-    "@angular/material": "^9.0.0",
-    "@angular/material-moment-adapter": "^9.0.0",
-    "@angular/platform-browser": "~9.1.0",
-    "@angular/platform-browser-dynamic": "~9.1.0",
-    "@angular/router": "~9.1.0",
+    "@angular/forms": "~12.0.1",
+    "@angular/material": "^12.0.1",
+    "@angular/material-moment-adapter": "^12.0.1",
+    "@angular/platform-browser": "~12.0.1",
+    "@angular/platform-browser-dynamic": "~12.0.1",
+    "@angular/router": "~12.0.1",
     "@ngx-translate/core": "^12.1.1",
-    "fp-ts": "^1.13.0",
-    "ngx-color-picker": "^10.0.1",
-    "polytype": "^0.11.0",
-    "lodash": "^4.17.19",
+    "fp-ts": "^2.11.5",
+    "ngx-color-picker": "^11.0.0",
+    "polytype": "^0.13.2",
+    "lodash": "^4.17.21",
     "rxjs": "~6.5.4",
-    "zone.js": "~0.10.2",
+    "zone.js": "~0.11.4",
     "uuid": "^8.3.2"
   },
   "dependencies": {
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.0"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.159"
